src/main/java/com/killeent/Array/ArraySearch.java
package com.killeent.Array;

import java.util.Comparator;

/**
 * Various search algorithms for arrays.
 *
 * @author Trevor Killeen (2014)
 */
public class ArraySearch {

    /**
     * Searches for the specified value in the array using binary search. Assumes that the array
     * is sorted. If the comparator's ordering is inconsistent with equals, this function
     * may behave strangely.
     *
     * @param array The array to search through. Should be sorted.
     * @param comparator Comparator to use when searching.
     * @return The index where the specified value is in the array, or -1 if the value
     * is not in the array.
     */
    public static <T> int binarySearch(T[] array, Comparator<? super T> comparator, T value) {
        int lo = 0;
        int hi = array.length;

        while (hi - lo > 0) {
            int mid = hi - ((hi - lo) / 2);
            int relation = comparator.compare(array[mid], value);
            if (relation < 0) {
                // a[mid] < value, search upper half of array
                lo = mid + 1;
            } else if (relation == 0) {
                // a[mid] == value, return index
                return mid;
            } else {
                // a[mid] > value, search lower half of array
                hi = mid;
            }
        }
        // not found
        return -1;
    }

}
